==> notebooks/1-preprocessing.ipynb <==
import pandas as pd
import numpy as np
import matplotlib.pyplot as plt
import seaborn as sns
from datasets import load_dataset, DatasetDict, Dataset
from datasets import Dataset as HFDataset
import nltk

# Download necessary NLTK resources

from nltk.corpus import stopwords
from nltk.tokenize import word_tokenize

import string


nltk.download('punkt')
nltk.download('stopwords')
nltk.download('wordnet')

ds = load_dataset('imdb')

ds_train = pd.DataFrame(ds['train']).head(10000)
ds_test = pd.DataFrame(ds['test']).head(10000)
ds_train.head()

# Function to remove punctuation from the text col
def remove_punctuation(text):
    text = "".join([char for char in text if char not in string.punctuation])

    return text

# Run the function on the datasets
ds_train['text_no_punct'] = ds_train['text'].apply(lambda x: remove_punctuation(x))
ds_test['text_no_punct'] = ds_test['text'].apply(lambda x: remove_punctuation(x))

import nltk
from nltk.tokenize import word_tokenize

def tokenize(text):
    # Convert text to lowercase and tokenize using nltk
    tokens = word_tokenize(text.lower())
    return tokens

# Run the tokenizer function on the df
nltk.download('punkt_tab')
ds_train['text_tokenized'] = ds_train['text_no_punct'].apply(lambda x: tokenize(x))
ds_test['text_tokenized'] = ds_test['text_no_punct'].apply(lambda x: tokenize(x))

ds_train.head()

# Calculate the length of each tokenized review
ds_train['text_length'] = ds_train['text_tokenized'].apply(len)

# Calculate minimum, maximum, and average number of words
min_length = ds_train['text_length'].min()
max_length = ds_train['text_length'].max()
avg_length = ds_train['text_length'].mean()

# Calculate the length of each tokenized review
ds_test['text_length'] = ds_test['text_tokenized'].apply(len)

# Calculate minimum, maximum, and average number of words
min_length = ds_test['text_length'].min()
max_length = ds_test['text_length'].max()
avg_length = ds_test['text_length'].mean()

# Define a function to rid of stop words in English
def remove_stopwords(tokens):    
    stop_words = stopwords.words('english')
    return [word for word in tokens if word.lower() not in stop_words]

# Run the stop words function on the df
ds_train['text_no_stop'] = ds_train['text_tokenized'].apply(remove_stopwords)
ds_test['text_no_stop'] = ds_test['text_tokenized'].apply(remove_stopwords)

ds_train.to_csv('/Users/blairjdaniel/lighthouse/lighthouse/LLM-Sentiment-Analysis-lhl/csv_files/train.csv')
ds_test.to_csv('/Users/blairjdaniel/lighthouse/lighthouse/LLM-Sentiment-Analysis-lhl/csv_files/test.csv')
